logging_s3: split validator tests into validators_test.go
